+++ src/store/Store.js
@@ -1,27 +1,13 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
-import * as getters from './getters'
-import * as mutations from './mutations'
+
+import todoApp from './modules/todoApp'
 
 Vue.use(Vuex);
 
-const storage = {
-    fetch() {
-        const arr = [];
-        if (localStorage.length > 0) {
-            for (let i = 0; i < localStorage.length; i++) {
-                if (localStorage.key(i) !== 'loglevel:webpack-dev-server') {
-                    arr.push(JSON.parse(localStorage.getItem(localStorage.key(i))));
-                }
-            }
-        }
-        return arr;
-    },
-};
 export const store = new Vuex.Store({
-    state: {
-        todoItems: storage.fetch() //관리하고 싶은 아이템을 등록한다?
+    modules: {
+        todoApp: todoApp
     },
-    getters: getters,
-    mutations: mutations
+
 });
