+++ 프로젝트구조화와모듈화.md
@@ -5,3 +5,12 @@ Vuex.Store({})  각각의 내용을 모듈화 시킨다면 !?
 ![image-20200810132608400](https://tva1.sinaimg.cn/large/007S8ZIlgy1ghllpfsh1qj30qb0b2djf.jpg)
 
 각각의 핼퍼 함수를 파일로 나눠서 저장함.
+
+
+
+브랜치 확인 - `vuex-with-helper-modules-with-properties`
+
+
+
+![image-20200810133507271](https://tva1.sinaimg.cn/large/007S8ZIlgy1ghllysezhzj30qc0e3783.jpg)
+
